Rescue timeout as Faraday::TimeoutError when net-http-persistent is simplified
(instead of `Faraday::ConnectionFailed`)

In v3 net-http-persistent rescue a lot of exceptions and retries them or re-raises them as Net::HTTP::Persistent::Error which is complicated and unnecessary.

When https://github.com/drbrain/net-http-persistent/pull/100 gets merged (or when using [net-http-persistent-retry](https://github.com/grosser/net-http-persistent) which is a fork of net-http-persistent with that PR applied) the error wrapping/retry goes away and it will start raising "normal" errors like `Net::OpenTimeout` which is a `Timeout::Error`.

/cc @drbrain @mislav @iMacTia